prelude: Also export `glib` and `gio`
This is a followup to "Dependencies are re-exported" from
https://gtk-rs.org/blog/2021/06/22/new-release.html
that we tried to follow in https://github.com/ostreedev/ostree-rs/pull/13

But now I'm trying to change `ostree-rs-ext` to be able to just do
`import ostree::prelude::*;` and be able to use e.g. `gio::Cancellable`,
but that didn't work without this explicit import of the toplevel
items too.